# Makes the PooledBlockAllocatorProvider a singleton