--- a/src/com/amazon/ion/impl/bin/PooledBlockAllocatorProvider.java
+++ b/src/com/amazon/ion/impl/bin/PooledBlockAllocatorProvider.java
@@ -20,8 +20,9 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ConcurrentMap;
 
 /**
- * A simple pooling implementation of {@link BlockAllocatorProvider} with a global thread-safe free block list
+ * A singleton implementation of {@link BlockAllocatorProvider} offering a thread-safe free block list
  * for each block size.
+ *
  * <p>
  * This implementation is thread-safe.
  */
@@ -76,11 +77,22 @@ import java.util.concurrent.ConcurrentMap;
         public void close() {}
     }
 
+    // A globally shared instance of the PooledBlockAllocatorProvider that is lazily initialized.
+    // This instance allows BlockAllocators to be re-used across instantiations of classes like
+    // the binary Ion writer, thereby avoiding costly array initializations.
+    private static PooledBlockAllocatorProvider instance;
     private final ConcurrentMap<Integer, BlockAllocator> allocators;
 
-    public PooledBlockAllocatorProvider()
+    private PooledBlockAllocatorProvider()
     {
         allocators = new ConcurrentHashMap<Integer, BlockAllocator>();
+    }
+
+    public static synchronized PooledBlockAllocatorProvider getInstance() {
+        if (instance == null) {
+            instance = new PooledBlockAllocatorProvider();
+        }
+        return instance;
     }
 
     @Override