--- a/src/com/amazon/ion/impl/bin/_PrivateIon_HashTrampoline.java
+++ b/src/com/amazon/ion/impl/bin/_PrivateIon_HashTrampoline.java
@@ -28,7 +28,7 @@ import java.io.IOException;
 @Deprecated
 public class _PrivateIon_HashTrampoline
 {
-    private static final PooledBlockAllocatorProvider ALLOCATOR_PROVIDER = new PooledBlockAllocatorProvider();
+    private static final PooledBlockAllocatorProvider ALLOCATOR_PROVIDER = PooledBlockAllocatorProvider.getInstance();
     
     public static IonWriter newIonWriter(ByteArrayOutputStream baos) throws IOException
     {