@almann 2021-05-24
> I wrote [a simple, unscientific benchmark](https://gist.github.com/zslayton/45194bfab526703e89e655a2eb8a7232) that creates a binary `IonWriter`, writes a value to a `ByteArrayOutputStream`, and then closes the `IonWriter` in a loop. The results below are the median of 5 runs each.

For a change like this, it might be worth writing a proper [JMH benchmark][1] to make sure we're not fooling ourselves.

[1]: https://github.com/openjdk/jmh