@almann 2021-05-24
> The provider should construct instances of the allocator and users should use the allocator, this changes the contract a bit--in other words do we have a problem in downstream code not creating this once using it everywhere?

Here is what the API documentation says for `BlockAllocatorProvider`:

https://github.com/amzn/ion-java/blob/45234885de653eae67db4d8d3ccb6bb85bb82fd9/src/com/amazon/ion/impl/bin/BlockAllocatorProvider.java#L18-L24

Doesn't this change violate the contract?