@tgregg 2021-05-24
I recommend using the [ion-java-benchmark-cli](https://github.com/amzn/ion-java-benchmark-cli) instead of writing a bespoke JMH benchmark. It's perfect for stuff like this.