@tgregg 2021-05-24
While we're in here, we should consider declaring the `PooledBlockAllocator` inner class `static`.