@almann 2021-05-24
> > The provider should construct instances of the allocator and users should use the allocator, this changes the contract a bit--in other words do we have a problem in downstream code not creating this once using it everywhere?
> 
> Here is what the API documentation says for `BlockAllocatorProvider`:
> 
> https://github.com/amzn/ion-java/blob/45234885de653eae67db4d8d3ccb6bb85bb82fd9/src/com/amazon/ion/impl/bin/BlockAllocatorProvider.java#L18-L24

Re-reading this a bunch, I think there is no technical violation here.  The *Provider* itself is the thread-safe thing, if we want to make the it a global singleton, I see no contract that would be violated by that.

Also, all of this is internal, so we should be okay here, but just surveying the usage and implications is a good thing.