@zslayton 2021-05-24
> I recommend using the ion-java-benchmark-cli instead of writing a bespoke JMH benchmark. It's perfect for stuff like this.

🤦 I should've used this, but I've already got the JMH code. Next time!

> While we're in here, we should consider declaring the PooledBlockAllocator inner class static.

Done.