@zslayton 2021-05-24
### JMH Benchmark results

Source can be found [here](https://gist.github.com/zslayton/849bf528c9ad4389c098099afc798a14).

**Before**
```
Benchmark                                 Score           Error   Units
time                                   1630.774 ±       135.671   ms/op
gc.alloc.rate                          7820.193 ±       492.074  MB/sec
gc.alloc.rate.norm              17549795353.600 ±    152386.604    B/op
gc.churn.G1_Eden_Space                 7652.005 ±       518.475  MB/sec
gc.churn.G1_Eden_Space.norm     17173158297.600 ± 841670600.565    B/op
gc.churn.G1_Old_Gen                       0.213 ±         0.307  MB/sec
gc.churn.G1_Old_Gen.norm             478707.200 ±    697129.834    B/op
gc.churn.G1_Survivor_Space                0.934 ±         0.059  MB/sec
gc.churn.G1_Survivor_Space.norm     2097152.000 ±         0.001    B/op
gc.count                                166.000                  counts
gc.time                                 299.000                      ms
```
**After**
```
Benchmark                                 Score           Error   Units
time                                    844.091 ±       124.800   ms/op
gc.alloc.rate                          3098.544 ±       299.204  MB/sec
gc.alloc.rate.norm               4390142753.600 ±     93107.441    B/op
gc.churn.G1_Eden_Space                 2833.914 ±       884.221  MB/sec
gc.churn.G1_Eden_Space.norm      4011432345.600 ± 913155271.731    B/op
gc.churn.G1_Old_Gen                       0.064 ±         0.120  MB/sec
gc.churn.G1_Old_Gen.norm              90323.200 ±    161209.039    B/op
gc.churn.G1_Survivor_Space                1.480 ±         0.143  MB/sec
gc.churn.G1_Survivor_Space.norm     2097152.000 ±         0.001    B/op
gc.count                                 87.000                  counts
gc.time                                 197.000                      ms
```